pyproject.toml
[project]
name = "body-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "firebase-admin>=6.7.0",
    "opencv-python>=4.11.0.86",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "scalar-fastapi>=1.0.3",
    "ultralytics>=8.3.107",
    "uvicorn>=0.34.0",
    "websockets>=15.0.1",
]
